refactor(core/query-builder): extract ColumnExpression into query-builder/expression (#405)
Pure move; re-exported from SelectQueryBuilder.ts.
